Day 17, part 1
Good god it's done. Interesting approach, lots of challenges, lots of
things to keep in mind. It's (as always is the case) been almost correct
for a while and the occasional missing '- k' really screwed me.

Insane the speed up in the algorithm for implementing a priority_queue
rather than searching each time for the next closest non-visited node. I
presumed it would be an improvement, but it just wouldn't run without
it, takes less than a second with it.