Day 20, part 2 incomplete
First attempt - always try part 2 using the same brute force as part 1
if you can. In this case, does not work and there is basically no
achievable optimisation - it has to be done mathematically, or I could
leave this running for a century and get my start post WWIV when I rule
the wasteland which is all that remains of civilisation.