Day 20, part 2
I've seen some complaints about this not being a great part, as it's
PSPACE-hard, there's unlikely to be a generalised solution as you may
get infinite cycles, etc. I think that's kind of true, but selective
input is generally part of these problems.

For me, the biggest issue was that you had to go peer into the input and
figure out something very particular about the structure. Now, that's
often a real life job you have to do, but it seems less in the spirit of
AoC, so I can understand the irritation.

Aside from that, basically reused the code from part 1, nice and easy.