Day 11, part 2
Really pleased with myself here. I had an inkling of this in the first
part, but it was quick enough not to need to bother. The insight I had
was that, since the question said "find another way to keep your worry
levels manageable", there must be some way to reduce the number being
passed on. The first guess I had was to find the lcm of the monkeys'
'test' numbers and set each worry level to itself modulo the lcm.

My reasoning was that this would keep encoded any factors of the test
numbers and therefore the tests would remain correct, which is all we
cared about, and that shifting by a number (adding) and multiplying
would make no difference to the result. Intuitively makes sense, thank
god I was right.