Day 16, part 2
Definitely need some optimisations, but got the correct answer out of
it. The issue is that this gives ~87000 combinations of valves, so the
hash comparison at the end runs over 7.5 billion times. Would need to
look at reducing the number of different combinations produced,
possibly, or some sort of heuristic to skip comparisons at the end.