Day 19, part 1
I'm not a fan purely cos there's no way for me to prove the heuristic,
but I cull all but the top 100,000 possibilities every minute. The
possibilities are ordered by most geodes and most obsidian and geode
robots - on the reasoning that this is the setup that will produce the
most future geodes and geode robots. I have no proof of this being
correct, except that it provided the correct answer in reasonable time.